Feat : Category 삭제시 포함된 메뉴가 없는지 체크 추가

+++ src/main/java/com/example/dp/domain/category/service/impl/CategoryServiceImpl.java
@@ -5,9 +5,11 @@ import com.example.dp.domain.category.dto.response.CategoryResponseDto;
 import com.example.dp.domain.category.entity.Category;
 import com.example.dp.domain.category.exception.CategoryErrorCode;
 import com.example.dp.domain.category.exception.ExistsCategoryTypeException;
+import com.example.dp.domain.category.exception.ForbiddenDeleteCategory;
 import com.example.dp.domain.category.exception.NotFoundCategoryException;
 import com.example.dp.domain.category.repository.CategoryRepository;
 import com.example.dp.domain.category.service.CategoryService;
+import com.example.dp.domain.menucategory.repository.MenuCategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,6 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final MenuCategoryRepository menuCategoryRepository;
 
     @Override
     public CategoryResponseDto createCategory(final CategoryRequestDto requestDto) {
@@ -46,6 +49,11 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void deleteCategory(final Long categoryId) {
         Category category = findCategory(categoryId);
+
+        if (menuCategoryRepository.existsByCategoryId(categoryId)) {
+            throw new ForbiddenDeleteCategory(CategoryErrorCode.FORBIDEN_DELETE);
+        }
+
         categoryRepository.delete(category);
     }
